feat: add tasks to the extension
