--- extension/ui/pages/ProjectMainPage.tsx
@@ -1,30 +1,0 @@
-import { BlockedActionsProvider } from "@app/components/assistant/conversation/BlockedActionsProvider";
-import { GenerationContextProvider } from "@app/components/assistant/conversation/GenerationContextProvider";
-import { PodPage } from "@app/components/pages/pod/PodPage";
-import { useAuth } from "@app/lib/auth/AuthContext";
-import { useSpaceInfo } from "@app/lib/swr/spaces";
-import { ConversationLayout } from "@extension/ui/components/conversation/ConversationLayout";
-import { ExtensionInputBarProvider } from "@extension/ui/components/conversation/ExtensionInputBarProvider";
-import { useParams } from "react-router-dom";
-
-export const ProjectMainPage = () => {
-  const { workspace } = useAuth();
-  const { podId } = useParams<{ podId: string }>();
-
-  const { spaceInfo } = useSpaceInfo({
-    workspaceId: workspace.sId,
-    spaceId: podId ?? null,
-  });
-
-  return (
-    <ConversationLayout title={spaceInfo?.name ?? ""}>
-      <BlockedActionsProvider owner={workspace}>
-        <GenerationContextProvider>
-          <ExtensionInputBarProvider workspace={workspace}>
-            <PodPage />
-          </ExtensionInputBarProvider>
-        </GenerationContextProvider>
-      </BlockedActionsProvider>
-    </ConversationLayout>
-  );
-};
