+++ front/components/pages/pod/PodPage.tsx
@@ -1,42 +1,15 @@
-import type { TaskOwnerFilter } from "@app/components/assistant/conversation/space/conversations/project_tasks/projectTasksListScope";
-import { ManageUsersPanel } from "@app/components/assistant/conversation/space/ManageUsersPanel";
-import { PodConversationsTab } from "@app/components/pod/conversation/PodConversationsTab";
-import { PodFilesTab } from "@app/components/pod/files/PodFilesTab";
 import { PodHeaderActions } from "@app/components/pod/PodHeaderActions";
-import { PodSettingsTab } from "@app/components/pod/settings/PodSettingsTab";
-import { PodTasksTab } from "@app/components/pod/tasks/PodTasksTab";
-import {
-  type PodConversationListFilter,
-  usePodConversations,
-} from "@app/hooks/conversations/usePodConversations";
+import { PodPageContent } from "@app/components/pod/PodPageContent";
 import { useActivePodId } from "@app/hooks/useActivePodId";
-import { useCreateConversationWithMessage } from "@app/hooks/useCreateConversationWithMessage";
-import { useSendNotification } from "@app/hooks/useNotification";
 import { useScopedPodUiPreferences } from "@app/hooks/useScopedUIPreferences";
 import {
   DEFAULT_POD_UI_PREFERENCES,
   type PodTab,
   usePodTabs,
 } from "@app/hooks/useSpaceProjectTabs";
-import { getLightAgentMessageFromAgentMessage } from "@app/lib/api/assistant/citations";
 import { useAuth, useWorkspace } from "@app/lib/auth/AuthContext";
-import { useClientType } from "@app/lib/context/clientType";
-import type { DustError } from "@app/lib/error";
-import { useAppRouter } from "@app/lib/platform";
 import { useSpaceInfo } from "@app/lib/swr/spaces";
 import { useIsMobile } from "@app/lib/swr/useIsMobile";
-import { getConversationRoute } from "@app/lib/utils/router";
-import type { LightConversationType } from "@app/types/assistant/conversation";
-import {
-  isAgentMessageType,
-  isUserMessageType,
-} from "@app/types/assistant/conversation";
-import type { RichMention } from "@app/types/assistant/mentions";
-import { toMentionType } from "@app/types/assistant/mentions";
-import type { ContentFragmentsType } from "@app/types/content_fragment";
-import type { Result } from "@app/types/shared/result";
-import { Err, Ok } from "@app/types/shared/result";
-import { removeNulls } from "@app/types/shared/utils/general";
 import {
   ChatBubbleLeftRightIcon,
   CheckCircleIcon,
@@ -44,19 +17,14 @@ import {
   FolderIcon,
   Spinner,
   Tabs,
-  TabsContent,
   TabsList,
   TabsTrigger,
 } from "@dust-tt/sparkle";
-import { useCallback, useState } from "react";
 
 export function PodPage() {
   const owner = useWorkspace();
   const { user } = useAuth();
-  const clientType = useClientType();
-  const router = useAppRouter();
   const podId = useActivePodId();
-  const sendNotification = useSendNotification();
 
   const {
     spaceInfo: podInfo,
@@ -69,39 +37,13 @@ export function PodPage() {
     includeAllMembers: true,
   });
 
-  const createConversationWithMessage = useCreateConversationWithMessage({
-    owner,
-    user,
-  });
-
   const { value: podUiPreferences, setValue: setPodUiPreferences } =
     useScopedPodUiPreferences({
       scope: "podUi",
       resourceId: podId,
       defaultValue: DEFAULT_POD_UI_PREFERENCES,
     });
-  const isSingleMemberPod = !!podInfo && podInfo.members.length === 1;
-  const conversationFilter: PodConversationListFilter = isSingleMemberPod
-    ? "all"
-    : podUiPreferences.conversationsFilter;
 
-  const {
-    conversations,
-    isConversationsLoading,
-    mutateConversations,
-    hasMore,
-    isEmpty: isPodEmpty,
-    loadMore,
-    isLoadingMore,
-  } = usePodConversations({
-    workspaceId: owner.sId,
-    podId,
-    filter: conversationFilter,
-  });
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [_planLimitReached, setPlanLimitReached] = useState(false);
-  const [isInvitePanelOpen, setIsInvitePanelOpen] = useState(false);
   const compactPodTabs = useIsMobile();
 
   const { currentTab, handleTabChange } = usePodTabs({
@@ -110,136 +52,6 @@ export function PodPage() {
     setPodUiPreferences,
   });
 
-  const handleConversationFilterChange = (
-    filter: PodConversationListFilter
-  ) => {
-    setPodUiPreferences({
-      ...podUiPreferences,
-      conversationsFilter: filter,
-    });
-  };
-
-  const handleTaskOwnerFilterChange = (tasksOwnerFilter: TaskOwnerFilter) => {
-    setPodUiPreferences({
-      ...podUiPreferences,
-      tasksOwnerFilter,
-    });
-  };
-
-  const handleConversationCreation = useCallback(
-    async (
-      input: string,
-      mentions: RichMention[],
-      contentFragments: ContentFragmentsType,
-      selectedMCPServerViewIds?: string[]
-    ): Promise<Result<undefined, DustError>> => {
-      if (isSubmitting) {
-        return new Err({
-          code: "internal_error",
-          name: "AlreadySubmitting",
-          message: "Already submitting",
-        });
-      }
-
-      setIsSubmitting(true);
-
-      const conversationRes = await createConversationWithMessage({
-        messageData: {
-          input,
-          mentions: mentions.map(toMentionType),
-          contentFragments,
-          selectedMCPServerViewIds,
-        },
-        spaceId: podId,
-      });
-
-      setIsSubmitting(false);
-
-      if (conversationRes.isErr()) {
-        if (conversationRes.error.type === "plan_limit_reached_error") {
-          setPlanLimitReached(true);
-        } else {
-          sendNotification({
-            title: conversationRes.error.title,
-            description: conversationRes.error.message,
-            type: "error",
-          });
-        }
-
-        return new Err({
-          code: "internal_error",
-          name: conversationRes.error.title,
-          message: conversationRes.error.message,
-        });
-      } else {
-        // Navigate to the new conversation
-        await router.push(
-          getConversationRoute(owner.sId, conversationRes.value.sId),
-          undefined,
-          { shallow: true }
-        );
-
-        // Converting to LightConversationType as createConversationWithMessage returns a ConversationType.
-        const lightConversation: LightConversationType = {
-          ...conversationRes.value,
-          content: removeNulls(
-            conversationRes.value.content.map((v) => {
-              const lastVersion = v[v.length - 1];
-              if (isUserMessageType(lastVersion)) {
-                return {
-                  ...lastVersion,
-                  // We don't really care about content fragments for light conversations in the UI.
-                  contentFragments: [],
-                };
-              }
-              if (isAgentMessageType(lastVersion)) {
-                return getLightAgentMessageFromAgentMessage(lastVersion);
-              }
-              return null;
-            })
-          ),
-        };
-
-        // Update the conversations list (prepend new conversation to first page)
-        await mutateConversations(
-          (currentData) => {
-            if (!currentData || currentData.length === 0) {
-              return [
-                {
-                  conversations: [lightConversation],
-                  hasMore: false,
-                  lastValue: null,
-                  isEmpty: false,
-                },
-              ];
-            }
-            const [firstPage, ...restPages] = currentData;
-            return [
-              {
-                ...firstPage,
-                conversations: [lightConversation, ...firstPage.conversations],
-              },
-              ...restPages,
-            ];
-          },
-          { revalidate: false }
-        );
-
-        return new Ok(undefined);
-      }
-    },
-    [
-      isSubmitting,
-      owner,
-      podId,
-      sendNotification,
-      router,
-      mutateConversations,
-      createConversationWithMessage,
-    ]
-  );
-
-  // Show loading state while fetching space info
   if (isPodsInfoLoading) {
     return (
       <div className="flex h-full w-full items-center justify-center">
@@ -248,7 +60,6 @@ export function PodPage() {
     );
   }
 
-  // Handle pod not found or access denied
   if (podInfoError || !podInfo) {
     return (
       <div className="flex h-full w-full items-center justify-center">
@@ -259,29 +70,6 @@ export function PodPage() {
             have access to it.
           </p>
         </div>
-      </div>
-    );
-  }
-
-  if (clientType === "extension") {
-    return (
-      <div className="flex h-full w-full flex-col">
-        <PodConversationsTab
-          owner={owner}
-          user={user}
-          conversations={conversations}
-          isConversationsLoading={isConversationsLoading}
-          hasMore={hasMore}
-          loadMore={loadMore}
-          isLoadingMore={isLoadingMore}
-          podInfo={podInfo}
-          isPodEmpty={isPodEmpty}
-          conversationFilter={conversationFilter}
-          onConversationFilterChange={handleConversationFilterChange}
-          onSubmit={handleConversationCreation}
-          onOpenMembersPanel={() => setIsInvitePanelOpen(true)}
-          onNavigateToTasks={() => handleTabChange("tasks")}
-        />
       </div>
     );
   }
@@ -335,56 +123,14 @@ export function PodPage() {
             )}
         </div>
 
-        <TabsContent value="conversations">
-          <PodConversationsTab
-            owner={owner}
-            user={user}
-            conversations={conversations}
-            isConversationsLoading={isConversationsLoading}
-            hasMore={hasMore}
-            loadMore={loadMore}
-            isLoadingMore={isLoadingMore}
-            podInfo={podInfo}
-            isPodEmpty={isPodEmpty}
-            conversationFilter={conversationFilter}
-            onConversationFilterChange={handleConversationFilterChange}
-            onSubmit={handleConversationCreation}
-            onOpenMembersPanel={() => setIsInvitePanelOpen(true)}
-            onNavigateToTasks={() => handleTabChange("tasks")}
-          />
-        </TabsContent>
-
-        <TabsContent value="files">
-          <PodFilesTab owner={owner} pod={podInfo} />
-        </TabsContent>
-
-        <TabsContent value="tasks">
-          <PodTasksTab
-            owner={owner}
-            podInfo={podInfo}
-            taskOwnerFilter={podUiPreferences.tasksOwnerFilter}
-            onTaskOwnerFilterChange={handleTaskOwnerFilterChange}
-          />
-        </TabsContent>
-
-        <TabsContent value="settings">
-          <PodSettingsTab
-            key={podId}
-            owner={owner}
-            pod={podInfo}
-            onOpenMembersPanel={() => setIsInvitePanelOpen(true)}
-          />
-        </TabsContent>
+        <PodPageContent
+          podInfo={podInfo}
+          onTabChange={handleTabChange}
+          podUiPreferences={podUiPreferences}
+          setPodUiPreferences={setPodUiPreferences}
+          mutatePodInfo={mutatePodInfo}
+        />
       </Tabs>
-      <ManageUsersPanel
-        isOpen={isInvitePanelOpen}
-        setIsOpen={setIsInvitePanelOpen}
-        owner={owner}
-        mode="space-members"
-        space={podInfo}
-        currentProjectMembers={podInfo.members}
-        onSuccess={() => mutatePodInfo()}
-      />
     </div>
   );
 }
